reorganized into src directory
