+++ main.go
@@ -1,14 +1,10 @@
 package main
 
 import (
-	"encoding/binary"
-	"errors"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
-	"sort"
-	"strings"
+    ft "github.com/osteensco/fastTravelCLI/src"
 )
 
 // TODO
@@ -38,354 +34,6 @@ import (
 
 
 
-func verifyInput(s string) (bool, error) {
-    switch strings.ToLower(s) {
-        case "y":
-            return true, nil
-        case "n":
-            return false, nil
-        default:
-            return false, errors.New(fmt.Sprintf("%v is not a valid response, please type y/n",s))
-    } 
-}
-
-func printMap(hashmap map[string]string) {
-
-    fmt.Println("")
-
-    keys := make([]string, 0, len(hashmap))
-    for k := range hashmap {
-        keys = append(keys, k)
-    }
-
-    sort.Strings(keys)
-
-    for i := range keys {
-
-        fmt.Printf("%v: %v\n", keys[i], hashmap[keys[i]])
-    
-    }
-
-    fmt.Println("")
-
-}
-
-func passCmd(args []string) ([]string, error) {
-    
-    cmd := args[1]
-
-    _, ok := availCmds[cmd]
-    if !ok {
-        return nil, errors.New(fmt.Sprintf("%v is not a valid command, use 'ft help' for valid commands", cmd))
-    }
-    switch cmd {
-        case "ls":
-            break
-        case "help":
-            break
-        case "rn":
-            if len(args) <= 3 {
-                return nil, errors.New(fmt.Sprintf("Insufficient args provided %v, usage: ft rn <key> <newKey>", args[1:]))
-            }
-        default:
-            if len(args) <= 2 {
-                return nil, errors.New(fmt.Sprintf("Insufficient args provided %v, usage: ft <command> <path/key>", args[1:]))
-            }
-    }
-
-	return args[1:], nil
-
-}
-
-
-
-func readMap(file *os.File) map[string]string {
-
-    pathMap := make(map[string]string)
-    
-
-    fileInfo, err := file.Stat()
-    if err != nil {
-        fmt.Println("Error getting file info: ", err)
-        os.Exit(1)
-    }
-
-    // take file size in bytes and make a buffer of that size
-    size := fileInfo.Size()
-    buff := make([]byte, size)
-
-    // read entire file into memory
-    _, err = file.Read(buff)
-    if err != nil {
-        fmt.Println("Error reading file into buffer: ", err)
-        os.Exit(1)
-    }
-
-    // key length integer should always fit in 1 byte
-    var keyLen uint8
-    // value length integer should always fit in 2 bytes
-    var valLen uint16
-    // sliding pointer to navigate buffer
-    var offset uint
-    
-    // iterate through buffer and deserialize
-    for offset < uint(len(buff)) {
-
-        // read length of key, use length to read in key, adjust offset
-        // simple type conversion since length is only 1 byte and not a []byte
-        keyLen = uint8(buff[offset])
-        offset++
-        kl := uint(keyLen)
-        keyBytes := buff[offset:offset+kl]
-        offset += kl
-        
-        // read length of value, use length to read in value, adjust offset
-        // length contained in 2 bytes, nedd to convert []byte to a uint16 value
-        valLen = binary.LittleEndian.Uint16(buff[offset:offset+2])
-        offset += 2
-        vl := uint(valLen)
-        valBytes := buff[offset:offset+vl]
-        offset += vl
-        // add key-value to map
-        pathMap[string(keyBytes)] = string(valBytes)
-
-    }
-
-	return pathMap
-
-}
-
-func sanitizeDir(path string) string {
-	distro := os.Getenv("WSL_DISTRO_NAME")
-	if len(distro) == 0 {
-
-    	prefix := "/mnt/"
-        drive := strings.Split(path, ":")[0]
-        path = strings.Replace(path, drive, strings.ToLower(drive), 1)
-        path = strings.Replace(path, ":", "", 1)
-    	path = strings.Replace(path, "\\", "/", -1)
-    	path = prefix + path
-
-	}
-    return path
-}
-
-func ensureData(filepath string) *os.File {
-
-    file, err := os.OpenFile(filepath, os.O_CREATE|os.O_RDWR, 0644)
-    if err != nil {
-        fmt.Println("Error opening file: ", err)
-        os.Exit(1)
-    }
-
-    return file
-
-}
-
-
-
-
-func dataUpdate(hashmap map[string]string, file *os.File) {
-    
-    var buffer []byte
-    for key, val := range hashmap {
-
-        keyBytes := []byte(key)
-        valBytes := []byte(val)
-        
-        keyLen := make([]byte, 1)
-        keyLen[0] = byte(uint8(len(keyBytes)))
-        
-        valLen := make([]byte,2)
-        binary.LittleEndian.PutUint16(valLen, uint16(len(valBytes)))
-
-        // create an array of array of bytes for optimal concatenation
-        allBytes := [][]byte{keyLen, keyBytes, valLen, valBytes}
-
-        // get the length for key-value pair to allocate memory
-        var pairLen int
-        for _, s := range allBytes {
-            pairLen += len(s)
-        }
-        // create new slice and append all []byte from allBytes
-        pair := make([]byte, pairLen)
-        var i int
-        for _, s := range allBytes {
-            i += copy(pair[i:], s)
-        }
-        // append completed pair to buffer []byte
-        allPairs := make([]byte, len(buffer)+len(pair))
-        copy(allPairs, buffer)
-        copy(allPairs[len(buffer):], pair)
-        buffer = allPairs
-    }
-    
-    err := file.Truncate(0) 
-    if err != nil {
-        fmt.Println("Error truncating file: ", err)
-        os.Exit(1)
-    }
-    
-    _, err = file.Seek(0,0) 
-    if err != nil {
-        fmt.Println("Error seeking to beginning of file: ", err)
-        os.Exit(1)
-    }
-
-    _, err = file.Write(buffer)
-    if err != nil {
-        fmt.Println("Error writing contents of buffer to file: ", err)
-        os.Exit(1)
-    }
-}
-
-
-func changeDirectory(data cmdArgs) {
-
-	if len(data.allPaths) == 0 {
-		fmt.Printf("No fast travel locations set, set locations by navigating to desired destination directory and using 'ft set <key>' ")
-		os.Exit(1)
-	}
-
-	p := data.allPaths[data.cmd[1]]
-    path := sanitizeDir(p)
-    fmt.Println(path)
-
-}
-
-func setDirectoryVar(data cmdArgs) {
-    
-    key := data.cmd[1]
-	path, err := os.Getwd()
-	if err != nil {
-		fmt.Println("Error:", err)
-		os.Exit(1)
-	}
-
-    for _,v := range data.allPaths {
-        if path == v {
-            fmt.Printf("Path '%v' already exists", path)
-            return
-        }
-    }
-
-    _, ok := data.allPaths[key]
-    if !ok {
-	    data.allPaths[key] = path
-        dataUpdate(data.allPaths, data.file)
-        fmt.Printf("Added destination %v", data.cmd[1] )
-    } else {
-        fmt.Printf("Key '%v' already exists", key)
-    }
-
-
-}
-
-func displayAllPaths(data cmdArgs) {
-
-    printMap(data.allPaths)
-
-}
-
-func removeKey(data cmdArgs) {
-    
-    var res string
-    key := data.cmd[1]
-
-    fmt.Printf("Are you sure you want to remove the key '%v'?", key)
-    _, err := fmt.Fscan(data.rdr, &res)
-    if err != nil {
-		fmt.Println("Error:", err)
-		os.Exit(1) 
-    }
-    if rm, err := verifyInput(res); rm {
-        fmt.Printf("Aborted removal of key %v", key)
-        os.Exit(1)
-    } else {
-        if err != nil {
-            fmt.Println("Error:", err)
-            os.Exit(1) 
-        }
-    }
-    delete(data.allPaths, key)
-    dataUpdate(data.allPaths, data.file)
-    fmt.Printf("Removed '%v' destination", key)
-    
-}
-
-func renameKey(data cmdArgs) {
-    
-    originalKey := data.cmd[1]
-    newKey := data.cmd[2]
-    
-    _, ok := data.allPaths[newKey]
-    if ok {
-        fmt.Printf("Key %v already exists, please choose something else.", newKey)
-        os.Exit(1)
-    }
-    path, ok := data.allPaths[originalKey]
-    if !ok {
-        err := errors.New(fmt.Sprintf("Cannot rename %v, key does not exist. Run 'ft ls' to see all keys.", originalKey))
-        fmt.Println(err)
-        os.Exit(1)
-    }
-    
-    var res string
-    
-    fmt.Printf("Are you sure you want to rename the key '%v'?", newKey)
-    _, err := fmt.Fscan(data.rdr, &res)
-    if err != nil {
-		fmt.Println("Error:", err)
-		os.Exit(1) 
-    }
-    if rm, err := verifyInput(res); rm {
-        fmt.Printf("Aborted renaming of key %v", newKey)
-        os.Exit(1)
-    } else {
-        if err != nil {
-            fmt.Println("Error:", err)
-            os.Exit(1) 
-        }
-    }
-    delete(data.allPaths, originalKey)
-    data.allPaths[newKey] = path 
-
-    dataUpdate(data.allPaths, data.file)
-    fmt.Printf("%v renamed to %v", originalKey, newKey)
-
-}
-
-func showHelp(data cmdArgs) {
-    
-    printMap(cmdDesc)
-
-}
-
-type cmdArgs struct {
-	cmd      []string
-	allPaths map[string]string
-    file *os.File
-    rdr io.Reader
-}
-
-// map of available commands
-var availCmds = map[string]func(data cmdArgs){
-	"to":  changeDirectory,
-	"set": setDirectoryVar,
-	"ls":  displayAllPaths,
-    "rm": removeKey,
-    "rn": renameKey,
-    "help": showHelp,
-}
-
-var cmdDesc = map[string]string {
-    "to": "change directory to provided key's path - Usage: ft to [key]",
-    "set": "set current directory path to provided key - Usage: ft set [key]",
-    "ls": "display all current key value pairs - Usage: ft ls",
-    "rm": "deletes provided key - Usage: ft rm [key]",
-    "rn": "renames key to new key - Usage: ft rn [key] [new key]",
-    "help": "you are here :) - Usage: ft help",
-}
